+++ src/application.ts
@@ -1,18 +1,18 @@
-import { createServer, ServerResponse } from "http";
+import { createServer } from "http";
 import { HttpServer } from "./types/http";
 import { Router } from "./router/route";
+import { Handlers } from "./types";
 export interface Application {
   listen: (port: number, onSuccess: () => void) => void;
   get: (path: string, handlers: Handlers) => void;
 }
 
-export type Handlers = (req: any, res: TypeExpressResponse) => void;
 export class TypeExpress implements Application {
   server: HttpServer;
-  router: Router
+  router: Router;
 
   constructor() {
-    this.router = new Router()
+    this.router = new Router();
     this.server = createServer((req, res) => {
       this.router.createRoute(req, res);
     });
@@ -27,15 +27,5 @@ export class TypeExpress implements Application {
       path,
       handlers,
     });
-  }
-}
-
-export class TypeExpressResponse {
-  constructor(private response: ServerResponse) {}
-
-  send(message: string): void {
-    this.response.setHeader("Content-Type", "text/plain");
-    this.response.write(message);
-    this.response.end();
   }
 }
