+++ src/request.ts
@@ -1,0 +1,6 @@
+import { IncomingMessage } from "http";
+
+export class TypeExpressRequest {
+  params!: string;
+  constructor(private request: IncomingMessage) {}
+}
