+++ src/types/index.ts
@@ -1,0 +1,7 @@
+import { TypeExpressRequest } from "../request";
+import { TypeExpressResponse } from "../response";
+
+export type Handlers = (
+  req: TypeExpressRequest,
+  res: TypeExpressResponse
+) => void;
